Show tabs in a light gray color.
For file types where the tabs are used for indentation, the tabs
are shown with a pipe (|) symbol.

For file types where the tabs are used for tabulation, the tabs
are shown with a greater than (>) symbol.

This is opposite to what is logical, but it looks better since
when it's used for indentation, the tab always appears as the
first character in the logical cell, so they all line up.  When
they're used for tabulation, the tab usually appears after some
other characters, so they don't line up.  (If only Vim let me show
a list character as the *last* character rather than the first.)