Try not showing the current command name in the title.
I plan to only open one shell per project, cd to that project directory,
then run tmux or vim or emacs as a shell.  In that scenario, the command
name is uninteresting.  This will also encourage me to cd to the right
place.  This makes the title more useful, and makes spawning processes
from the "parent" "shell" just do the right thing.