Print two times in the zsh rprompt.
The first time is the time the prompt was printed.

The second time is the current time, updated every second, until a command is
run.

This way, I'll have the current time (the second time). And a way to easily see
roughly how long each command took to run (the second time for the previous
prompt minus the first time for the current prompt).