i3: Don't focus new windows.
It's confusing to have the focus be somewhere other than where the mouse cursor
is when using focus_follows_mouse. You create a new window, work in it, then
come back later, and start working on the window where the mouse pointer is, but
the input goes to a different window.