vim: Disable mouse mode, line numbers, and the sign column
I wanted to make the clipboard work properly with mouse mode. The best option
seemed to be to enable OSC 52 support to set the primary selection whenever Vim
set the visual selection, then use the terminal's copy command to copy the
PRIMARY selection to the X11 CLIPBOARD.

It turns out:

1) most terminals don't support OSC 52's "p" parameter to set the primary
selection
2) Vim doesn't have any Visual mode events like VisualSelect or VisualLeave that
could be used to trigger this

For now, give up on mouse mode, and disable the things like the gutter and line
numbers that get in the way of copying using the terminal's own selection.